BOL-375 refactor: file 관련 model(in data module) package 분리
