+++ data/src/main/java/com/yapp/bol/data/datasource/file/FileDataSource.kt
@@ -1,6 +1,6 @@
 package com.yapp.bol.data.datasource.file
 
-import com.yapp.bol.data.model.group.response.ImageFileUploadResponse
+import com.yapp.bol.data.model.file.ImageFileUploadResponse
 import com.yapp.bol.domain.model.ApiResult
 import kotlinx.coroutines.flow.Flow
 import java.io.File
